move definitions to super class to reuse same method
